Add Jenkinsfile for auto build discovery
Prior to this commit we were required to manaully create Jenkins jobs
for building and testing this repository. This commit adds a shell of a
Jenkinsfile that enumerates the various stages in the pipeline and
successfully executes the Checkout and Build stages. The Test and
Release stages are merely placeholders.