Change Jenkinsfile to out-of-source build
Prior to this commit Jenkins would build the binaries in the source tree
alongside the source files, generally acknowledged to be a bad practice.
This commit alters the Jenkins build to create those binaries in a
separate `build` directory so that the binaries can be referenced by
unit and acceptance tests in a standard location.