Separate filters into plugin pattern.
Before this commit, the command line parsing in transcode.cpp was
complicated by having many transcoder features processed in the same
place and together.  Filters, encoders, and the toml processing have
been moved to separate translation units to make each feature easier
to read.  In the filter case, they are now plugins rather like the
encoders, so we can add specialized or premium filters (like a sifter)
later on just like an encoder.  The command line processing is still
necessarily complicated for such a tool, but it is now better documented
in transcode.cpp.