illustrate example extract-to-function refactoring
Prior to this commit all the functionality was contained within `main`
which made it difficult to ascertain what, at a high-level, the
transcoder was doing. This commit pulls out a couple of isolated
functions in order to illustrate how, at a very gross level, how
functionality can be extracted into functions with descriptive names.