Rearrange catch_main()
Prior to this commit, catch_main() in transcode.cpp executing linearly
with some of the options defined inline.  This commit moved some of the
self contained blocks into separate functions, in an attempt to make the
necessarily complicated parser easier to read.  Note that most of the
options are already set in their own functions in plugins as part of the plugin
interface, and never were in catch_main().  Whether this commit makes the
function catch_main() easier to read or not is subjective matter, and
does diverge somewhat from the style established by the reference
documentation of boost::program_options.