Re-enable unit tests
Prior to this commit, the unit tests required Mettle to be installed on
the build machine.  This merge will build libmettle.so and provide the
headers within the transcoder distribution to avoid this external dependency, and allowing the tests to build and run again.

There are also whitespace removal edits, as my editor touches each file
after configuring it to remove trailing whitespace.