Remove LFS to workaround Jenkins issue
Prior to this commit, the plog artifacts used for testing were checked
into LFS.  The combination of github/LFS/Jenkins is somehow broken, so
this commit works around the problem by checking in the binary files
into git, not git-lfs.  We would rather revert back to LFS when the problem
is fixed.