Feature/ps 1463 jenkins mettle (#20)
* Adapt CMakeLists and README.md for out-of-PolySync build

Prior to this commit, CMakeLists.txt assumed the presence of conan, as
used inside of PolySync.  Furthermore, there was no installation
instructions.

* Coax cmake to find clang instead of gcc

Prior to this commit, cmake defaulted to gcc.  On Ubuntu 16.04,
transcoder requires clang.

* Fix conanfile dependencies

Prior to this commit, boost was building static only, which does not
work for the transcoder plugins.

* Move behavioral test executable location

Prior to this commit, the output files were in a pre-conan place.  This
merge makes behave find polysync-transcode in the new location.

* Move include directory for conan

Prior to this commit, the deps/ directory was not found by the new conan
setup.  This commit changes CMakeLists.txt to always look at the source
directory for deps.

* Rename "deps" directory to "vendor"

Prior to this commit, vendored dependencies lived in deps, which turns
out was also used by Jenkins for something else, with the result that the vendored libraries
were blown away in Jenkins.  This commit just renames the directory to
avoid confusion with our Jenkins pattern.

* Change inclusions of cpptoml.h from deps/ to vendor/

Prior to this commit, cpptoml.h lived in ./deps.  This name has changed
to ./vendor.

* Patch Jenkinsfile to erase garbage after broken builds

Prior to this commit, the Jenkins build would leave junk around after a
(broken) build.  This is a company-wide patch to clean up Jenkins no
matter what happens.