Feature/ps 1463 where is share (#23)
* Print error message when no TOML descriptions are found

Prior to this commit, the transcoder would happily run without warning as long as some share/ path was provided, even if the path did not actually yield any type descriptions.  This was confusing for new users if they had not yet learned to configure a path, or if the default pathfinding failed.  This commit adds the warning message, and broke out print_vector.hpp to get the IO support needed.

* Add install path as a default resource location for TOML and plugins.

Prior to this commit, the transcoder missed a good guess at where the TOML and plugins might be, because it did not know the install path.  This commit adds CMAKE_INSTALL_PREFIX/share to the default search path for resources.