Move artifacts directory
Prior to this commit, the example plogs lived in features/artifacts.
These are useful enough for lots of experiments that digging into
features/ all the time is tedious, and these support files deserve a top
level directory made by this commit.  Plus this should fix the behave
tests in Jenkins.