Set rpath
Prior to this commit, the shared library libtranscode.so had to install
to a system library path, which failed even on /usr/local/lib because
apparently that is not in the default Ubuntu ld path.  This commit uses
CMake to set up the rpath so the shared libs are always found (on the
executable and also the plugins).