Check input file is actually a plog (#24)
* Check input file is actually a plog

Prior to this commit, the decoder attempted to decode any given file
assuming that it was a plog.  If the file was not a plog, a crash was
likely.  This commit adds a "magic number" checker as a rough check that
the input file actually is a plog.  If we revise the plog format, we
will add an actual magic number to make this test failsafe, but for now
we just do a heuristic check on the plog version numbers.

* Reset stream position always and check for all zeros in magic check

Prior to this commit, the magic number checker would pass all  zeros (probably not a plog) and would not reset the stream always if the check failed.