Unspecialize decoder from plog
Prior to this commit, the decoder only ever decoded plog files.  It
turns out that it is sufficiently generic that it can also decode rosbag
with only a little work.  This commit is the first cut at removing plog
dependency from the decoder.  This commit passes both unit and
acceptance tests.