Switch detector location in TOML
Prior to this commit, the detector was associated with the precursor
type.  This was a drag when adding new types, because you had to
pollute, for instance, core.toml with detectors for velodyne and ibeo
types.  This would make it hard to develop new TOML without writing
/usr/local/share/polysync-transcode/core.toml.  This commit moves the
detector to the precursor type where it belongs.  Now you reference core
types like ps_byte_array_msg in velodyne.toml, not the other way around.