Fix crashing bug in the detector catalog on invalid TOML
Prior to this commit, it was possible to crash the transcoder by writing
a non-integer value in the detector fields.  It crashed ultimately
because std::stroul() could not parse the string into an integer.  This
commit catches the error and throws a polysync error with context, so
the user can figure out what is wrong with his TOML.