Add descriptor types for Bit and BitSkip
Prior to this commit, bitfields such as those used in CAN could not be
described because all previous types respected byte boundaries.  This
commit adds the description types for bit level fields.  They do not yet
parse or work; this is the first of several commits to on the way to
supporting CAN.