Update the README.md with first-use information
Prior to this commit the `README.md` file was missing key information about
the Transcoder tool and where resources were located within the repo.

This commit adds a first pass to the README file to include information about
what the tool can do, when it is used, how to install it and where examples
can be found.